getting odd values with float vectors / image pixel values

+++ include/henson_sonar/coarseDM.h
@@ -10,10 +10,10 @@ public:
   CoarseDM();
   // Eigen::Matrix<int, 1681, 169> dictionaryMatrix(int pixelX, int pixelY, cv::Mat imgTarget, cv::Mat imgRef);
 
-  Eigen::VectorXi getGamma(int x, int y, cv::Mat img);
-  Eigen::Matrix<int, Dynamic, Dynamic> dictionaryMatrix(int pixelX, int pixelY, cv::Mat imgTarget, cv::Mat imgRef);
+  Eigen::VectorXf getGamma(int x, int y, cv::Mat img);
+  Eigen::Matrix<float, Dynamic, Dynamic> dictionaryMatrix(int pixelX, int pixelY, cv::Mat imgTarget, cv::Mat imgRef);
   // Eigen::Matrix<int, 2, 169> getTargetErrorOMP(Matrix<int, 1681, 169> dictA, VectorXi targetX);
-  Eigen::Matrix<float, Dynamic, Dynamic> getTargetErrorOMP(const Matrix<int, Dynamic, Dynamic>& dictA, const VectorXi& targetX);
+  Eigen::Matrix<float, Dynamic, Dynamic> getTargetErrorOMP(const Matrix<float, Dynamic, Dynamic>& dictA, const VectorXf& targetX);
 private:
 
 };
